פיצול קוד בניית הסרטונים
פיצול לקוד יצירת אודיו ווידאו נפרדים
