+++ scripts/build-v2-multi-languages.py
@@ -4,7 +4,6 @@ import os
 import random
 import re
 import numpy as np
-from google.cloud import texttospeech
 from moviepy.editor import *
 from PIL import Image, ImageDraw, ImageFont
 import arabic_reshaper
@@ -13,9 +12,7 @@ from concurrent.futures import ThreadPoolExecutor, as_completed
 import tempfile
 import logging
 from functools import lru_cache
-
-# הגדרת נתיב למפתח ה-API (ודאו שהקובץ JSON נמצא במיקום המתאים)
-os.environ["GOOGLE_APPLICATION_CREDENTIALS"] = r"C:\Users\me\OneDrive\וידאו\מפתחות גישה\youtube-channel-440320-fe17f0f0a940.json"
+from audio_creator import AudioCreator
 
 # הגדרת רמת הלוגינג
 logging.basicConfig(level=logging.INFO, format='%(asctime)s - %(levelname)s - %(message)s')
@@ -241,71 +238,6 @@ class ImageCreator:
         self.cache[cache_key] = img
         return img
 
-class AudioCreator:
-    def __init__(self, temp_dir, lang_settings):
-        self.temp_dir = temp_dir
-        self.lang_settings = lang_settings
-        self.executor = ThreadPoolExecutor(max_workers=THREADS)
-        self.client = texttospeech.TextToSpeechClient()
-
-    def create_audio_task(self, text, lang, slow=False):
-        try:
-            voice_config = self.lang_settings.get(lang, self.lang_settings['en']).get('voice', None)
-            if voice_config is None:
-                logging.error(f"לא נמצאו הגדרות קול עבור שפה: {lang}")
-                raise ValueError(f"לא נמצאו הגדרות קול עבור שפה: {lang}")
-
-            synthesis_input = texttospeech.SynthesisInput(text=text)
-            voice_params = texttospeech.VoiceSelectionParams(
-                language_code=voice_config['language_code'],
-                name=voice_config['name']
-            )
-            audio_config = texttospeech.AudioConfig(
-                audio_encoding=texttospeech.AudioEncoding.MP3,
-                speaking_rate=0.70 if slow else 0.95
-            )
-
-            response = self.client.synthesize_speech(
-                input=synthesis_input,
-                voice=voice_params,
-                audio_config=audio_config
-            )
-
-            with tempfile.NamedTemporaryFile(delete=False, suffix='.mp3', dir=self.temp_dir.name) as tmp_file:
-                tmp_file.write(response.audio_content)
-                return tmp_file.name
-
-        except ValueError as e:
-            logging.error(f"שגיאה ביצירת אודיו עבור הטקסט: '{text}' בשפה: '{lang}'. פרטים: {e}")
-            raise
-
-    def create_audios(self, tasks):
-        futures = {}
-        for task in tasks:
-            if len(task) == 3:
-                text, lang, slow = task
-                future = self.executor.submit(self.create_audio_task, text, lang, slow)
-            else:
-                text, lang = task
-                future = self.executor.submit(self.create_audio_task, text, lang, False)
-            futures[future] = task
-
-        results = {}
-        for future in as_completed(futures):
-            task = futures[future]
-            try:
-                audio_path = future.result()
-                results[tuple(task)] = audio_path
-                logging.info(
-                    f"אודיו נוצר עבור: '{task[0]}' בשפה: '{task[1]}' עם slow={'True' if len(task) == 3 and task[2] else 'False'}"
-                )
-            except Exception as e:
-                logging.error(f"שגיאה ביצירת אודיו עבור: '{task[0]}' בשפה: '{task[1]}'. פרטים: {e}")
-        return results
-
-    def shutdown(self):
-        self.executor.shutdown(wait=True)
-
 class VideoCreator:
     def __init__(self, file_manager, image_creator, audio_creator, style_definitions, lang_settings):
         self.file_manager = file_manager
@@ -423,7 +355,6 @@ class VideoCreator:
         )
         return clip_intro
 
-
     def create_outro(self, lang_code):
         outro_data = self.lang_settings.get(lang_code, self.lang_settings['en']).get('outro', None)
         if outro_data is None:
@@ -445,7 +376,6 @@ class VideoAssembler:
         self.video_creator = VideoCreator(file_manager, image_creator, audio_creator, style_definitions, lang_settings)
         self.lang_code = None
         self.lang_settings = lang_settings
-
 
     def assemble_level_video(self, level, output_dir, thumbnails_dir, lang_code):
         self.lang_code = lang_code
@@ -620,7 +550,7 @@ def main():
             lang_settings = json.load(f)
 
         image_creator = ImageCreator(styles=style_definitions)
-        audio_creator = AudioCreator(file_manager.temp_dir, lang_settings)
+        audio_creator = AudioCreator(file_manager.temp_dir, lang_settings, THREADS)
         video_assembler = VideoAssembler(file_manager, image_creator, audio_creator, style_definitions, lang_settings)
 
         for level in data['levels']:
